Make the PREVIEW text be hidden when hovering
This PREVIEW text makes sense when choosing a theme on an existing site but not in MLB signup flow.

On MLB blogs, the team-specific styles are implemented as theme_options on each of the partner-mlb-fan,
partner-mlb-modern and partner-mlb-retro themes.
Current preview logic does not seem to support thumbnails from alternatives offered by a theme as theme_options.